refactor: move `PaymentContext` into sub-directory
