+++ packages/shared/src/components/plus/PlusDesktop.tsx
@@ -2,8 +2,8 @@ import type { ReactElement } from 'react';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { useRouter } from 'next/router';
 import dynamic from 'next/dynamic';
-import type { OpenCheckoutFn } from '../../contexts/PaymentContext';
-import { usePaymentContext } from '../../contexts/PaymentContext';
+import type { OpenCheckoutFn } from '../../contexts/payment/PaymentContext';
+import { usePaymentContext } from '../../contexts/payment/PaymentContext';
 import { PlusInfo } from './PlusInfo';
 import { PlusCheckoutContainer } from './PlusCheckoutContainer';
 import { useGiftUserContext } from './GiftUserContext';
