+++ packages/shared/src/components/plus/PlusIOS.tsx
@@ -4,8 +4,8 @@ import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import { useQueryClient } from '@tanstack/react-query';
 import { PlusInfo } from './PlusInfo';
-import type { OpenCheckoutFn } from '../../contexts/PaymentContext';
-import { usePaymentContext } from '../../contexts/PaymentContext';
+import type { OpenCheckoutFn } from '../../contexts/payment/PaymentContext';
+import { usePaymentContext } from '../../contexts/payment/PaymentContext';
 import type { CommonPlusPageProps } from './common';
 import { promisifyEventListener } from '../../lib/func';
 import { webappUrl } from '../../lib/constants';
